% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_and_render.R
\name{diff_and_render}
\alias{diff_and_render}
\title{Compare Two Tabular Datasets}
\usage{
diff_and_render(a, b, title = NULL, outfile = NULL, summary = TRUE,
  ignore_whitespace = FALSE, never_show_order = TRUE,
  unchanged_context = 0L, ..., verbose = TRUE)
}
\arguments{
\item{a}{`tibble`, `data.frame`, or path to file}

\item{b}{`tibble`, `data.frame`, or path to file}

\item{title}{rendered in HTML}

\item{outfile}{if not `NULL`, then write to this file}

\item{summary}{see help for [daff::render_diff()]}

\item{ignore_whitespace}{see help for [daff::diff_data()]}

\item{never_show_order}{see help for [daff::diff_data()]}

\item{unchanged_context}{see help for [daff::diff_data()]}

\item{...}{further arguments to [daff::diff_data()]}

\item{verbose}{display informative messages}
}
\value{
a `diff` object, invisibly
}
\description{
`diff_and_render()` uses the `daff` package to compare two tabular datasets.
The difference is rendered as HTML, which is also (optionally) written to a file.
}
