% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_csv.R
\name{diff_csv}
\alias{diff_csv}
\title{Diff two CSV files}
\usage{
diff_csv(file1, file2, include = everything(), exclude = NULL, ...)
}
\description{
Diff two CSV files
}
\note{
migrated from `inventory` package, 2019-12-12
}
